Adjusted post-processing of Eigenvalue decomposition.
